# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "SberTask",
        "contact": {}
    },
    "host": "localhost:3003",
    "paths": {
        "/task": {
            "post": {
                "description": "Создает новую задачу",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Создать задачу",
                "parameters": [
                    {
                        "description": "Данные для создания задачи",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/task.CreateTask"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/task.Task"
                        }
                    }
                }
            }
        },
        "/task/{id}": {
            "get": {
                "description": "Получает задачу по заданному идентификатору",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Получить задачу по идентификатору",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/task.Task"
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет существующую задачу",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Обновить задачу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления задачи",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/task.Task"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/task.Task"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет задачу по заданному идентификатору",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Удалить задачу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Обновляет часть данных существующей задачи",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Частичное обновление задачи",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Часть данных для обновления задачи, строки могут быть пустыми",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/task.PartiallyUpdateTask"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/task.Task"
                        }
                    }
                }
            }
        },
        "/tasks": {
            "get": {
                "description": "Получает список всех задач",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Получить все задачи",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/task.Task"
                            }
                        }
                    }
                }
            }
        },
        "/tasks/date": {
            "post": {
                "description": "Получает список всех доступных задач по заданной дате",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Получить все задачи по определенной дате",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Запрос на получение задач с определенной датой и статусом",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/task.Task"
                            }
                        }
                    }
                }
            }
        },
        "/tasks/status": {
            "post": {
                "description": "Получает список всех задач с заданным статусом",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Получить все задачи с определенным статусом",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Запрос на получение задач с определенным статусом",
                        "name": "status",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/task.Task"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "task.CreateTask": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string",
                    "example": "2023-09-22T09:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Описание новой задачи"
                },
                "status": {
                    "type": "boolean",
                    "example": false
                },
                "title": {
                    "type": "string",
                    "example": "Новая задача"
                }
            }
        },
        "task.PartiallyUpdateTask": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string",
                    "example": "Обновленная дата 2023-09-21T12:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Обновленное Описание задачи 1"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "boolean",
                    "example": false
                },
                "title": {
                    "type": "string",
                    "example": "Обновленная Задача 1"
                }
            }
        },
        "task.Task": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string",
                    "example": "2023-09-21T12:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Описание задачи 1"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "boolean",
                    "example": false
                },
                "title": {
                    "type": "string",
                    "example": "Задача 1"
                }
            }
        }
    }
}